+++ src/parser/mrt/mod.rs
@@ -3,12 +3,12 @@ Parse MRT header and content, provides [BgpElem][BgpElem] struct for per-prefix 
 
 [BgpElem]: crate::BgpElem
 */
+pub mod messages;
 pub mod mrt_elem;
 pub mod mrt_header;
-pub mod mrt_message;
 pub mod mrt_record;
 
-pub use mrt_message::bgp4mp::parse_bgp4mp;
-pub use mrt_message::table_dump_message::parse_table_dump_message;
-pub use mrt_message::table_dump_v2_message::parse_table_dump_v2_message;
+pub use messages::bgp4mp::parse_bgp4mp;
+pub use messages::table_dump_message::parse_table_dump_message;
+pub use messages::table_dump_v2_message::parse_table_dump_v2_message;
 pub use mrt_record::parse_mrt_record;
